requirements.txt
python_dotenv
accelerate
transformers==4.40.2
#conda env config vars set PYTORCH_ENABLE_MPS_FALLBACK=1
#conda activate <test-env>
#pip3 install torch torchvision
torch
torchvision
chardet
#conda install pytorch torchvision -c pytorch